requirements.txt
# Core dependencies
httpx==0.27.0
pyyaml==6.0.1
jinja2==3.1.3

# Deployment (optional)
boto3==1.34.34

# Date/time handling
python-dateutil==2.8.2
